+++ spec/factories/integration_tokens.rb
@@ -1,0 +1,7 @@
+FactoryBot.define do
+  factory :integration_token do
+    association :created_by, factory: :user
+    name { "fan_tools" }
+    scopes { ["fan_tools.read"] }
+  end
+end